--- src/AssemblyInfo.cs
@@ -1,1 +1,0 @@
-﻿[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Identifiable.Test")]
